+++ src/main/java/org/openlmis/stockmanagement/i18n/MessageKeys.java
@@ -192,6 +192,9 @@ public abstract class MessageKeys {
   public static final String ERROR_LINE_ITEM_REASON_ID_NOT_FOUND = ERROR_PREFIX
       + ".lineItem.reason.id.notFound";
 
+  public static final String ERROR_LINE_ITEM_REASON_TAGS_INVALID = ERROR_PREFIX
+      + ".tags.invalid";
+
   //source destination configuration
   public static final String ERROR_ORGANIZATION_NAME_MISSING = ERROR_PREFIX
       + ".organization.name.missing";