+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 New functionality added in a backwards-compatible manner:
 * [OLMIS-4599](https://openlmis.atlassian.net/browse/OLMIS-4599): Add ability to retrieve a stock card reason by id
+* [OLMIS-4598](https://openlmis.atlassian.net/browse/OLMIS-4598): Add tags to StockCardLineItemReason
 
 3.0.0 / 2018-04-24
 ==================